+++ hashtagmap-admin/src/main/resources/application.yml
@@ -1,7 +1,4 @@
 spring:
-  mvc:
-    view:
-      suffix: .html
   resources:
     static-locations:
       - file:hashtagmap-admin/src/main/resources/static
