+++ hashtagmap-common/src/main/java/com/songpapeople/hashtagmap/response/CustomResponse.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 // Todo 순서 컨벤션 확인
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class CustomResponse<T> {
