+++ hashtagmap-kakao-scheduler/src/main/java/com/songpapeople/hashtagmap/scheduler/exception/KakaoSchedulerException.java
@@ -3,7 +3,11 @@ package com.songpapeople.hashtagmap.scheduler.exception;
 import com.songpapeople.hashtagmap.exception.HashtagMapException;
 
 public class KakaoSchedulerException extends HashtagMapException {
-    public KakaoSchedulerException(KakaoSchedulerExceptionStatus exceptionStatus) {
-        super(exceptionStatus.getMessage(), exceptionStatus.getStatusCode());
+    public KakaoSchedulerException(final KakaoSchedulerErrorCode kakaoSchedulerErrorCode) {
+        super(kakaoSchedulerErrorCode.getMessage(), kakaoSchedulerErrorCode.getCode());
+    }
+
+    public KakaoSchedulerException(final KakaoSchedulerErrorCode kakaoSchedulerErrorCode, final String detailMessage) {
+        super(kakaoSchedulerErrorCode.getMessage(), kakaoSchedulerErrorCode.getCode(), detailMessage);
     }
 }
